fix: restore signals before deferred cleanup
